requirements.txt
Django==5.2.4
django-tailwind==4.1.0
supabase==2.3.4
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-allauth==0.60.1
Pillow==10.2.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0 